% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_rasters.R
\name{binary_rasters}
\alias{binary_rasters}
\title{Species-specific threshold maps}
\usage{
binary_rasters(x, y)
}
\arguments{
\item{x}{vector or matrix}

\item{y}{raster layers in a stack, brick or list}
}
\value{
A map, histogram and calculated area in km2
}
\description{
Classify predicted species distribution maps into binary raster and visualize
}
